Move utils tests to integration tests to make setup and cleanup shareable [ci skip]
