+++ src/utils.rs
@@ -74,110 +74,3 @@ pub fn split_header_from_contents(content: &[u8]) -> Result<(&[u8], &[u8]), anyh
     let body = split_content.next().context("Getting body")?;
     Ok((header, body))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn setup() -> PathBuf {
-        let test_dir_name = uuid::Uuid::new_v4().to_string();
-        let test_dir = [".test", &test_dir_name].iter().collect::<PathBuf>();
-        if !test_dir.exists() {
-            fs::create_dir_all(&test_dir).unwrap();
-        }
-
-        test_dir
-    }
-
-    fn cleanup(path: PathBuf) {
-        fs::remove_dir_all(path).unwrap();
-    }
-
-    #[test]
-    fn split_header_from_contents_returns_header_and_body_success() {
-        let content = b"blob 5\0hello";
-        let (header, body) = split_header_from_contents(content).unwrap();
-        assert_eq!(header, b"blob 5");
-        assert_eq!(body, b"hello");
-    }
-
-    #[test]
-    fn split_header_from_body_error_no_null_byte() {
-        let content = b"blob 5hello";
-        let result = split_header_from_contents(content);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_create_header_success() {
-        let object_type = ObjectType::Blob;
-        let file = b"file content";
-        let header = create_header(&object_type, file);
-        assert_eq!(header, b"blob 12\0");
-    }
-
-    #[test]
-    fn test_split_header_from_contents_success() {
-        let content = b"blob 5\0hello";
-        let (header, body) = split_header_from_contents(content).unwrap();
-        assert_eq!(header, b"blob 5");
-        assert_eq!(body, b"hello");
-    }
-
-    #[test]
-    fn test_split_header_from_contents_error() {
-        let content = b"blob 5hello";
-        let result = split_header_from_contents(content);
-        assert!(result.is_err());
-    }
-
-    fn write_head_file(contents: &str) -> PathBuf {
-        let path = setup();
-        let head = path.join("not-git").join("HEAD");
-
-        fs::create_dir_all(head.parent().unwrap()).unwrap();
-        fs::write(head, contents).unwrap();
-
-        path
-    }
-
-    #[test]
-    fn test_get_head_ref_success() {
-        let branch_name: &str = "test_branch_name";
-        let path = write_head_file(&format!("ref: refs/heads/{}\n", branch_name));
-
-        let head_ref = get_head_ref(Some(&path)).unwrap();
-        assert_eq!(head_ref, branch_name);
-
-        cleanup(path);
-    }
-
-    #[test]
-    fn test_get_head_ref_error_no_file() {
-        let path = setup();
-
-        let result = get_head_ref(None);
-        assert!(result.is_err());
-
-        cleanup(path);
-    }
-
-    #[test]
-    fn test_get_head_ref_error_improper_format() {
-        let branch_name: &str = "test_branch_name";
-        let path = write_head_file(&format!("refs/heads/{}\n", branch_name));
-
-        let result = get_head_ref(None);
-        assert!(result.is_err());
-
-        cleanup(path);
-    }
-
-    // #[test]
-    // fn test_read_from_file_success() {
-    //     let content = read_from_file("/path/to/file.txt").unwrap();
-    //     assert_eq!(content, b"file content");
-    // }
-
-    // fn test_read_from_file_error_if_file_missing() {}
-}
